+++ documentation-site/src/pages/demos/space-shooter/_asteroid-spawner.system.ts
@@ -8,9 +8,9 @@ import {
 import { Random, Vector2 } from '@forge-game-engine/forge/math';
 import { addSpriteComponent } from '@forge-game-engine/forge/rendering';
 import {
-  addPhysicsBodyComponent,
-  CircleShape,
-  RigidBody,
+  addAabbComponent,
+  addColliderComponent,
+  CircleCollider,
 } from '@forge-game-engine/forge/physics';
 import {
   AsteroidSpawnerEcsComponent,
@@ -72,15 +72,10 @@ export const createAsteroidSpawnerEcsSystem = (
       const asteroidRadius =
         (sprite.width * asteroidScale + sprite.height * asteroidScale) / 4;
 
-      addPhysicsBodyComponent(world, asteroidEntity, {
-        physicsBody: new RigidBody({
-          shape: new CircleShape(asteroidRadius),
-          position: new Vector2(x, spawnerComponent.spawnY),
-          isStatic: false,
-          isSensor: true,
-        }),
-        isKinematic: true,
+      addColliderComponent(world, asteroidEntity, {
+        collider: new CircleCollider(asteroidRadius),
       });
+      addAabbComponent(world, asteroidEntity);
     }
   },
 });